Split ambiguity and repair-guidance issue-lint tests
