+++ src/supervisor/supervisor-diagnostics-issue-lint-ambiguity.test.ts
@@ -1,9 +1,38 @@
 import assert from "node:assert/strict";
 import test from "node:test";
-import {
-  createIssueLintFixture,
-} from "./supervisor-test-helpers";
+import { createIssueLintFixture } from "./supervisor-test-helpers";
 import { GitHubIssue } from "../core/types";
+
+test("issue lint does not flag concrete risky work as blocking ambiguity", async () => {
+  const { loadIssueLintReport } = await createIssueLintFixture();
+
+  const issue: GitHubIssue = {
+    number: 106,
+    title: "Rotate production auth tokens",
+    body: `## Summary
+Rotate the production auth token flow for service-to-service requests.
+
+## Scope
+- update auth token issuance for production services
+- keep rollout audit-friendly
+
+## Acceptance criteria
+- production authentication changes are fully implemented
+
+## Verification
+- npm test -- src/supervisor/supervisor-diagnostics-issue-lint-ambiguity.test.ts`,
+    createdAt: "2026-03-19T00:00:00Z",
+    updatedAt: "2026-03-19T00:00:00Z",
+    url: "https://example.test/issues/106",
+    state: "OPEN",
+  };
+
+  const report = await loadIssueLintReport(issue);
+
+  assert.match(report, /^issue=#106$/m);
+  assert.match(report, /^execution_ready=yes$/m);
+  assert.match(report, /^high_risk_blocking_ambiguity=none$/m);
+});
 
 test("issue lint reports high-risk blocking ambiguity distinctly", async () => {
   const { loadIssueLintReport } = await createIssueLintFixture();
@@ -22,7 +51,7 @@ Decide whether to keep the current production auth token flow or replace it befo
 - the operator confirms which auth flow should ship
 
 ## Verification
-- npm test -- src/supervisor/supervisor-diagnostics-issue-lint.test.ts`,
+- npm test -- src/supervisor/supervisor-diagnostics-issue-lint-ambiguity.test.ts`,
     createdAt: "2026-03-19T00:00:00Z",
     updatedAt: "2026-03-19T00:00:00Z",
     url: "https://example.test/issues/107",
