+++ src/supervisor/supervisor-diagnostics-issue-lint-repair-guidance.test.ts
@@ -3,7 +3,7 @@ import test from "node:test";
 import { createIssueLintFixture } from "./supervisor-test-helpers";
 import { GitHubIssue } from "../core/types";
 
-test("issue lint reports missing required execution-ready sections deterministically", async () => {
+test("issue lint reports repair guidance for missing required execution-ready sections", async () => {
   const { loadIssueLintReport } = await createIssueLintFixture();
 
   const issue: GitHubIssue = {
@@ -23,9 +23,21 @@ Issue lint should report missing sections.`,
   assert.match(report, /^execution_ready=no$/m);
   assert.match(report, /^missing_required=scope, acceptance criteria, verification$/m);
   assert.match(report, /^missing_recommended=depends on, execution order$/m);
+  assert.match(
+    report,
+    /^repair_guidance_1=Add a `## Scope` section with bullet points describing the in-scope work\.$/m,
+  );
+  assert.match(
+    report,
+    /^repair_guidance_2=Add a `## Acceptance criteria` section listing the observable completion checks\.$/m,
+  );
+  assert.match(
+    report,
+    /^repair_guidance_3=Add a `## Verification` section with the exact command, test file, or manual check to run\.$/m,
+  );
 });
 
-test("issue lint reports malformed and locally inconsistent scheduling metadata", async () => {
+test("issue lint reports repair guidance for malformed scheduling metadata", async () => {
   const { loadIssueLintReport } = await createIssueLintFixture();
 
   const issue: GitHubIssue = {
@@ -46,7 +58,7 @@ Parallelizable: Later
 - invalid metadata is called out clearly
 
 ## Verification
-- npx tsx --test src/supervisor/supervisor-diagnostics-issue-lint-metadata.test.ts`,
+- npx tsx --test src/supervisor/supervisor-diagnostics-issue-lint-repair-guidance.test.ts`,
     createdAt: "2026-03-19T00:00:00Z",
     updatedAt: "2026-03-19T00:00:00Z",
     url: "https://example.test/issues/104",
@@ -56,6 +68,8 @@ Parallelizable: Later
   const report = await loadIssueLintReport(issue);
 
   assert.match(report, /^issue=#104$/m);
-  assert.match(report, /^metadata_errors=part of references the issue itself; depends on contains malformed references: #oops; depends on references the issue itself; depends on repeats #105; execution order must be N of M with 1 <= N <= M; parallelizable must be Yes or No$/m);
-  assert.match(report, /^high_risk_blocking_ambiguity=none$/m);
+  assert.match(
+    report,
+    /^repair_guidance_1=Replace invalid scheduling metadata with valid `Part of: #<number>`, `Depends on: none|#<number>`, `Execution order: N of M`, and `Parallelizable: Yes|No` lines\.$/m,
+  );
 });
